+++ src/CSAppController.h
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2003,2006-2007, Bryan L Blackburn.  All rights reserved.
+ * Copyright © 2003,2006-2007,2011, Bryan L Blackburn.  All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -41,7 +41,7 @@ extern NSString * const CSDocumentPboardType;
 
 @interface CSAppController : NSObject
 {
-   int lastPBChangeCount;
+   NSInteger lastPBChangeCount;
    BOOL closeAllFromTimeout;
 
    IBOutlet NSMenuItem *editMenuSetCategory;
