+++ src/NSData_compress.m
@@ -42,11 +42,11 @@
 #import "NSData_compress.h"
 #include <zlib.h>
 
-const int NSDataCompressionLevelNone = Z_NO_COMPRESSION;
-const int NSDataCompressionLevelDefault = Z_DEFAULT_COMPRESSION;
-const int NSDataCompressionLevelLow = Z_BEST_SPEED;
-const int NSDataCompressionLevelMedium = 5;
-const int NSDataCompressionLevelHigh = Z_BEST_COMPRESSION;
+const NSInteger NSDataCompressionLevelNone = Z_NO_COMPRESSION;
+const NSInteger NSDataCompressionLevelDefault = Z_DEFAULT_COMPRESSION;
+const NSInteger NSDataCompressionLevelLow = Z_BEST_SPEED;
+const NSInteger NSDataCompressionLevelMedium = 5;
+const NSInteger NSDataCompressionLevelHigh = Z_BEST_COMPRESSION;
 
 
 // Localized strings
@@ -95,26 +95,29 @@ static BOOL compressLoggingEnabled = YES;
  * Compress the data at the given compression level; stores the original data
  * size at the end of the compressed data
  */
-- (NSMutableData *) compressedDataAtLevel:(int)level
+- (NSMutableData *) compressedDataAtLevel:(NSInteger)level
 {
    /*
     * zlib says to make sure the destination has 0.1% more + 12 bytes; last
     * additional bytes to store the original size (needed for uncompress)
     */
-   unsigned long bufferLength = ceil((float) [self length] * 1.001) + 12 + sizeof(unsigned);
+#warning 64BIT: Inspect use of unsigned long
+#warning 64BIT: Inspect use of sizeof
+   unsigned long bufferLength = ceil((CGFloat) [self length] * 1.001) + 12 + sizeof(NSUInteger);
    NSMutableData *newData = [NSMutableData dataWithLength:bufferLength];
    if(newData != nil)
    {
-      int zlibError = compress2([newData mutableBytes],
-                                &bufferLength,
-                                [self bytes],
-                                [self length],
-                                level);
+      NSInteger zlibError = compress2([newData mutableBytes],
+                                      &bufferLength,
+                                      [self bytes],
+                                      [self length],
+                                      level);
       if(zlibError == Z_OK)
       {
          // Add original size to the end of the buffer, written big-endian
-         *((unsigned *) ([newData mutableBytes] + bufferLength)) = NSSwapHostIntToBig([self length]);
-         [newData setLength:bufferLength + sizeof(unsigned)];
+         *((NSUInteger *) ([newData mutableBytes] + bufferLength)) = NSSwapHostIntToBig([self length]);
+#warning 64BIT: Inspect use of sizeof
+         [newData setLength:bufferLength + sizeof(NSUInteger)];
       }
       else
       {
@@ -139,8 +142,9 @@ static BOOL compressLoggingEnabled = YES;
    NSMutableData *newData = nil;
    if([self isCompressedFormat])
    {
-      unsigned int originalSize = NSSwapBigIntToHost(*((unsigned *) ([self bytes] + [self length] -
-                                                                     sizeof(unsigned))));
+      NSUInteger originalSize = NSSwapBigIntToHost(*((NSUInteger *) ([self bytes] + [self length] -
+#warning 64BIT: Inspect use of sizeof
+                                                                     sizeof(NSUInteger))));
       /*
        * In the rare circumstance that data which is not compressed happens to
        * pass the checks above, we need to deal with the possibility that there
@@ -166,11 +170,13 @@ static BOOL compressLoggingEnabled = YES;
       NS_ENDHANDLER
       if(newData != nil)
       {
+#warning 64BIT: Inspect use of unsigned long
          unsigned long outSize = originalSize;
-         int zlibError = uncompress([newData mutableBytes],
-                                    &outSize,
-                                    [self bytes],
-                                    [self length] - sizeof(unsigned));
+         NSInteger zlibError = uncompress([newData mutableBytes],
+                                          &outSize,
+                                          [self bytes],
+#warning 64BIT: Inspect use of sizeof
+                                          [self length] - sizeof(NSUInteger));
          if(zlibError != Z_OK)
          {
             [NSData logCompressMessage:NSLocalizedString(@"call to uncompress() failed: %d - %s", @""),
